CV2/support-gemini/requirements.txt
google-generativeai>=0.7.0
openai>=1.30.0
streamlit>=1.35.0
gspread>=6.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
gtts>=2.5.1
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
